Change Input component to MessageForm Use form submit instead of separated events for sending a message.
