--- src/components/Input/Input.scss
@@ -1,32 +1,0 @@
-.form {
-  display: flex;
-  justify-content: space-between;
-  border-top: 2px solid #d3d3d3;
-
-  input {
-    padding: 1.5rem;
-    width: 80%;
-    font-size: 15px;
-    border: none;
-    border-radius: 15px;
-
-    &:focus {
-      outline: none;
-    }
-  }
-}
-
-.btn-send {
-  display: inline-block;
-  padding: 20px;
-  background-color: #2979ff;
-  text-decoration: none;
-  text-transform: uppercase;
-  color: #fff;
-  border: none;
-  transition: all 0.3s;
-
-  &:hover {
-    background-color: #1f61d4;
-  }
-}
